refactored code pt1
