+++ back-end/wordlebot/wordlebot/settings.py
@@ -1,21 +1,10 @@
+from pathlib import Path
 import os
 
-"""
-Django settings for wordlebot project.
 
-Generated by 'django-admin startproject' using Django 4.1.3.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.1/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/4.1/ref/settings/
-"""
-
-from pathlib import Path
-
-# Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
+# SECRET_KEY = os.getenv('SECRET_KEY')
+# DEBUG = int(os.getenv('DEBUG', 0)) == 1
 
 
 # Quick-start development settings - unsuitable for production
@@ -27,7 +16,9 @@ SECRET_KEY = "django-insecure-oh6l9b^lq0l^m2q8ch%)=6s23p2=_g6rwl_s0by4-#@un7lb0i
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
 
-ALLOWED_HOSTS = []
+ALLOWED_HOSTS = [
+    'localhost',
+]
 
 CORS_ALLOWED_ORIGINS = [
     'http://localhost:3000',
