--- front-end/src/api/index.ts
@@ -1,6 +1,0 @@
-const axios = require("axios");
-
-export async function makeRequest(path: string) {
-  //sends request to back-end
-  return axios.get(path);
-}
